Added positional symmetry to input file handling.

+++ R/full_run.R
@@ -24,26 +24,23 @@ full_run <- function(yaml_fpath, return_output = T) {
                   "csv" = data.table::fread(instr$scores_file), 
                   "rds" = readRDS(instr$scores_file))
   
-  .data <- collect_all_layer_configurations(.data)
   
-  #.data |> #map(~ {.x@guideGIs}) |> 
-  #  str() |> print()
+  .make_symmetric <- "make_symmetric" %in% names(instr) && (isTRUE(instr$make_symmetric))
+  
+  .data <- collect_all_layer_configurations(.data, make_pos_agnostic = .make_symmetric)
   
   #if (F) {
-  
   #.data <- imap(.data, ~ {print(.y); compute_dupCorrelation(.x)})
   .data <- map(.data, compute_dupCorrelation)
   
   
   # store all GI objects
-  if ("output_directory" %in% names(instr) & instr$overwrite_output) {
+  if ("output_directory" %in% names(instr) && (isTRUE(instr$overwrite_output))) {
     saveRDS(.data, file.path(instr$output_directory, "all_GI_objects.rds"))
   }
-    
-    
+  
   .keep_all <- "keep_all_configurations" %in% names(instr) && (isTRUE(instr$keep_all_configurations))
-  print("Keep all:")
-  print(.keep_all)
+
   .data <- find_optimal_configuration(.data, keep_all = .keep_all)
